임성찬: 마이페이지 내모임, 내댓글 조회시 결과값 수정, 모임 수정, 삭제 기능 실행 조건 추가

+++ controllers/moims.js
@@ -219,6 +219,10 @@ const updateMoim = async (req, res, next) => {
     console.log('target 모임', targetMoim);
     console.log('target 모임의 lengh', targetMoim.length);
 
+    if( targetMoim.userId !== userId ) {
+      return next(new Error('모임의 작성자만 수정이 가능합니다.'))
+    }
+
     if (targetMoim.length > 0) {
       //2. update 실행
       await Moim.update(
@@ -266,6 +270,10 @@ const deleteMoim = async (req, res, next) => {
     }).catch((err) => {
       if (err) next(new Error('delete의 targetMoim find 중 db 에러'));
     });
+
+    if( targetMoim.userId !== userId ) {
+      return next(new Error('모임의 작성자만 삭제가 가능합니다.'))
+    }
 
     if (targetMoim.length > 0) {
       await Moim.destroy({
